+++ autosub/options.py
@@ -43,6 +43,7 @@ def get_cmd_parser():  # pylint: disable=too-many-statements
                  "when the option is not given at the command line.\n"
                  "\"(arg_num)\" means if the option is given,\n"
                  "the number of the arguments is required.\n"
+                 "Arguments *ARE* the things given behind the options.\n"
                  "Author: {author}\n"
                  "Email: {email}\n"
                  "Bug report: {homepage}\n").format(
@@ -388,13 +389,13 @@ def get_cmd_parser():  # pylint: disable=too-many-statements
         '-sw1', '--stop-words-1',
         metavar=_('words_delimited_by_space'),
         help=_("(Experimental)First set of Stop words to split two events. "
-               "(arg_num = 1) (default: %(default)s)"))
+               "(arg_num = 1)"))
 
     conversion_group.add_argument(
         '-sw2', '--stop-words-2',
         metavar=_('words_delimited_by_space'),
         help=_("(Experimental)Second set of Stop words to split two events. "
-               "(arg_num = 1) (default: %(default)s)"))
+               "(arg_num = 1)"))
 
     conversion_group.add_argument(
         '-ds', '--dont-split',
@@ -517,7 +518,7 @@ def get_cmd_parser():  # pylint: disable=too-many-statements
                "audio conversion command. "
                "\"[\", \"]\" are optional arguments "
                "meaning you can remove them. "
-               "\"{{\", \"}}\" are required arguments "
+               "\"{\", \"}\" are required arguments "
                "meaning you can't remove them. "
                "(arg_num = 1) (default: %(default)s)"))